requirements.txt
-i https://pypi.org/simple
certifi==2019.11.28
chardet==3.0.4
click==7.0
flask==1.1.1
future==0.18.2
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.3
line-bot-sdk==1.15.0
markupsafe==1.1.1
requests==2.22.0
urllib3==1.25.7
werkzeug==0.16.0
pixivpy==3.5.7
gunicorn